Tweaks

+++ include/bullet.h
@@ -5,14 +5,14 @@
 
 class Bullet : public Actor {
     protected:
-        int damage;
+        float damage;
         float timer = 5;
     public:
-        Bullet(double x, double y, int size, double speed, double direction, int damage);
+        Bullet(double x, double y, int size, double speed, double direction, float damage);
 
         int getTimer(void) { return this->timer; };
 
-        int getDamage(void) { return this->damage; };
+        float getDamage(void) { return this->damage; };
 
         void update(float delta) override;
 };
