+++ include/cheryl.h
@@ -24,7 +24,7 @@ public:
 
     void update(float delta) override;
 
-    void damage(int d) override;
+    void damage(float d) override;
 };
 
 #endif
