+++ src/range_weapon.cpp
@@ -9,7 +9,7 @@
 #include <math.h>
 #define PI 3.14159265
 
-RangeWeapon::RangeWeapon(double x, double y, double width, double height, int damage, float fireRate) :
+RangeWeapon::RangeWeapon(double x, double y, double width, double height, float damage, float fireRate) :
 	Item(ActorType::RANGEWEAPON, x, y, width, height, 1, false) {
 	this->x = x;
 	this->y = y;
